pyproject.toml
[project]
name = "job-agent"
version = "0.1.0"
description = "Agentic system for discovering high-confidence job opportunities"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Colin Murphy", email = "clnjmurph@gmail.com" }]

dependencies = [
    # Core
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "pyyaml>=6.0",
    # Database
    "sqlalchemy>=2.0",
    "psycopg[binary]>=3.1",
    "alembic>=1.13",
    "pgvector>=0.2",
    # Web
    "fastapi>=0.109",
    "uvicorn[standard]>=0.27",
    "httpx>=0.26",
    # LLM / AI
    "openai>=1.10",
    "langchain>=0.1",
    "langgraph>=0.0.20",
    "langfuse>=2.0",
    # Dashboard
    "streamlit>=1.30",
    # Utilities
    "beautifulsoup4>=4.12",
    "lxml>=5.0",
    "tenacity>=8.2",
    "structlog>=24.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.1",
    "pytest-asyncio>=0.23",
    "ruff>=0.2",
    "mypy>=1.8",
    "pre-commit>=3.6",
    "types-PyYAML",
    "types-beautifulsoup4",
]

[project.scripts]
job-agent = "pipelines.daily:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app", "collectors", "pipelines", "schemas"]

[tool.ruff]
target-version = "py311"
line-length = 88

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # function call in default argument
    "B009",   # getattr with constant attribute
]

[tool.ruff.lint.isort]
known-first-party = ["app", "collectors", "pipelines", "schemas"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = [
    "langfuse.*",
    "langgraph.*",
    "streamlit.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["evals"]
markers = [
    "integration: marks tests as integration tests (deselect with '-m \"not integration\"')",
]
